Refactor list notes
